feat: Add new tests and refactor
